Separate io functions Separate report generation Removed buttons, report gen to be done with script Reformatted for clarity Fixed bug where lmfi heatmaps combined cultures Include pool level stats
